refactor and add drag and drop
