+++ projects/ngx-flow/src/lib/drop.directive.spec.ts
@@ -1,0 +1,8 @@
+import { DropDirective } from './drop.directive';
+
+describe('FlowDropDirective', () => {
+  it('should create an instance', () => {
+    const directive = new DropDirective();
+    expect(directive).toBeTruthy();
+  });
+});
