+++ projects/ngx-flow/src/lib/flow.directive.spec.ts
@@ -1,0 +1,8 @@
+import { FlowDirective } from './flow.directive';
+
+describe('FlowDirective', () => {
+  it('should create an instance', () => {
+    const directive = new FlowDirective();
+    expect(directive).toBeTruthy();
+  });
+});
