+++ projects/ngx-flow/src/lib/ngx-flow.module.ts
@@ -1,10 +1,12 @@
 import { NgModule } from '@angular/core';
 import { ButtonDirective } from './button.directive';
 import { SrcDirective } from './src.directive';
+import { DropDirective } from './drop.directive';
+import { FlowDirective } from './flow.directive';
 
 @NgModule({
   imports: [],
-  declarations: [ButtonDirective, SrcDirective],
-  exports: [ButtonDirective, SrcDirective]
+  declarations: [ButtonDirective, SrcDirective, DropDirective, FlowDirective],
+  exports: [ButtonDirective, SrcDirective, DropDirective, FlowDirective]
 })
 export class NgxFlowModule {}
