+++ projects/ngx-flow/src/lib/src.directive.spec.ts
@@ -1,8 +1,8 @@
-import { FlowSrcDirective } from './src.directive';
+import { SrcDirective } from './src.directive';
 
 describe('FlowSrcDirective', () => {
   it('should create an instance', () => {
-    const directive = new FlowSrcDirective();
+    const directive = new SrcDirective();
     expect(directive).toBeTruthy();
   });
 });
