+++ projects/ngx-flow/src/public_api.ts
@@ -2,7 +2,9 @@
  * Public API Surface of ngx-flow
  */
 
-export * from './lib/button.directive';
+// export * from './lib/button.directive';
+// export * from './lib/drop.directive';
+// export * from './lib/flow.directive';
 export * from './lib/upload-state';
 export * from './lib/transfer';
 export * from './lib/ngx-flow.module';
